chore: pull some general pieces out of linux IO module
